* origin-lela/master: (64 commits)
  * Fix xml in tutorial.xml * Add argument Ring R to VectorUtils::appendEntries * Use R.assign in VectorUtils::appendEntries
  * AbstractElement now contains pointer back to ring * Implement AbstractElement copy-constructor and assignment-operator to invoke ring->refElement * Change AbstractElement destructor to invoke ring->unrefElement * Add refElement, unrefElement to RingInterface * Update tutorial-documentation to reflect this * Add elementWidth to RingInterface to return width in characters of a ring-element * Implement elementWidth in various rings * Use elementWidth in MatrixWriter::writePretty
  Remove unused lela/randiter/modular-balanced.h
  Remove test-generic-for-quad.h and use standard ring-tests for GF(2)
  * Fixes to make distcheck work * Change version-number to 0.1.0
  Make some tweaks to avoid compiler-warnings about tautological compares and to avoid unneeded operations
  * Add disambiguating specialisations of CompareSparseEntries::operator () * Add DefaultVectorTraits<std::vector<std::pair<...> > > * Make SimpleAccessor work properly in the case that one uses STL-vector of pairs
  * Add section to tutorial on creating subvectors, submatrices, transposed matrices * Add some notes on potential difficulties when compiling * Add some notes on forbidden type-combinations with BLAS-operations * Add warning about working with mutable sparse subvectors
  * Add one more note on difference between LELA and LinBox to README * Add comments on element-types for Modular
  * Add comments to README on the differences between LELA and LinBox * Add storage-type M4RITranspose and use it to represent the transpose of a M4RI-matrix
  Use 0xffffffff rather than 0xdeadbeef for noSource
  * Remove spurious typename's * MatrixReader::read: Call BLAS3::scal only after matrix has been resized
  * Set Endianness correctly based on M4RI-version * M4RIMatrix::RowIterator: Call make_row only on dereference * MatrixRawIterator<Dense01>: Fetch iterator for next row only on dereference
  * Fix typo in blas-check.m4 * Use argument -std=c99 when checking whether M4RI-version is new
  * Remove all mentions of lapack from blas-check.m4 * Update documentation-strings for blas and gmp-checks * Fix some bugs in _dot<ZpModule> * Add more test-cases to test-blas-zp-module * Fix output of BLAS3-tests so that they don't refer to MatrixDomain * Tweak output of dot-product-test
  * Fix include-directives slightly so that we can call BLAS3::scal inside MatrixReader::read * MatrixReader::read: initialise matrix to 0 before reading * Don't warn in M4RI-module if we have already included level3.h -- no longer needed * Tweak output in test-blas-zp-module
  * Fix calls to cblas in transpose case * Use M.TWM for modules when passing through to _gemm
  * Fix DenseMatrix::RawIterator so that it works properly for submatrices * Make MatrixRawIterator<Dense> writable
  * Fix output of testgerConsistency * Fix coefficients passed to _gemv in _gemv<ZpModule> * Fix arguments to cblas_{s|d}gemv * Fix arguments to _gemm in _gemm<ZpModule> * Always do the first iteration in _gemm<ZpModule>
  * Add new test-blas-zp-module for testing ZpModule * Fix construction of const submatrix of a TransposeMatrix * Add new MatrixStorageType DenseTranspose * BLASModule now supports transpose of matrices and passes them correctly to CBLAS * Fix typo in _scal<BLASModule<double> > * Don't crash when block_size is 0 in ZpModule * Fix calls to reduce in ZpModule * Replace instances of DenseMatrix<float/double> with generic Matrix in BLASModule-implementations * Fix _gemv calls in ZpModule * ModularTraits<float>::DoubleFatElement should be float, not double * Add AllModules<TypeWrapperRing<Element> > TWM to ZpModule to be used when passing calls through
  ...

Conflicts:
	lela/blas/Makefile.am
	lela/blas/old.level1-modular.tcc
	lela/blas/old.level2-modular.tcc
	lela/randiter/Makefile.am
	lela/ring/Makefile.am
	lela/ring/interface.h
	macros/Makefile.am
	tests/Makefile.am
	tests/test-blas-generic-module.C
	tests/test-gauss-jordan.C
	tests/test-generic-for-quad.h
	tests/test-strassen-winograd.C

ADD: LIBPOLYS_HOME path in config.h
TODO: fix _newly_ wrong interface...

+++ tests/test-gauss-jordan.C
@@ -113,11 +113,12 @@ int main (int argc, char **argv)
 
 	parseArguments (argc, argv, args);
    
-        typedef MyModular<uint32> Ring;
+        typedef MyModular<float> Ring;
         typedef Ring::Element Element;
    
 	Ring GFq (q);
-	GF2 gf2;
+
+        GF2 gf2;
 
 	commentator.setBriefReportParameters (Commentator::OUTPUT_CONSOLE, false, false, false);
 	commentator.getMessageClass (INTERNAL_DESCRIPTION).setMaxDepth (5);
